@@ -8,6 +8,9 @@
     "start": "webpack serve --mode development",
     "docs": "webpack --mode development",
     "release": "standard-version",
+    "releasePatch": "npm run release -- --release-as patch",
+    "releaseMinor": "npm run release -- --release-as minor",
+    "releaseMajor": "npm run release -- --release-as major",
     "prepublishOnly": "npm run build",
     "eslint": "eslint {src,demo}\"/**/*.\"{js,jsx}",
     "prettier": "prettier {src,demo}\"/**/*.\"{js,jsx}",
@@ -27,40 +30,43 @@
   "author": "João Vitor <joaoassisb@gmail.com>",
   "license": "MIT",
   "devDependencies": {
-    "@babel/core": "^7.12.10",
-    "@babel/plugin-transform-runtime": "^7.12.10",
-    "@babel/preset-env": "^7.12.11",
-    "@babel/preset-react": "^7.12.10",
-    "@babel/runtime": "^7.12.5",
-    "@rollup/plugin-babel": "^5.2.2",
-    "@rollup/plugin-node-resolve": "^11.0.1",
-    "babel-jest": "^26.6.3",
-    "babel-loader": "^8.2.2",
-    "eslint": "^7.16.0",
-    "eslint-import-resolver-webpack": "^0.13.0",
-    "eslint-plugin-import": "^2.22.1",
-    "eslint-plugin-react": "^7.21.5",
-    "eslint-plugin-react-hooks": "^4.2.0",
-    "jest": "^26.6.3",
-    "js-var-type": "^0.2.6",
-    "lodash-es": "^4.17.20",
-    "prettier": "^2.2.1",
-    "prop-types": "^15.7.2",
-    "react": "^17.0.1",
-    "react-dom": "^17.0.1",
-    "react-hot-loader": "^4.13.0",
-    "rollup": "^2.35.1",
-    "standard-version": "^9.0.0",
-    "webpack": "^5.33.2",
-    "webpack-cli": "^4.6.0",
-    "webpack-dev-server": "^3.11.2"
+    "@babel/core": "7.26.0",
+    "@babel/plugin-transform-runtime": "7.25.9",
+    "@babel/preset-env": "7.26.0",
+    "@babel/preset-react": "7.25.9",
+    "@babel/runtime": "7.26.0",
+    "@eslint/compat": "1.2.2",
+    "@eslint/eslintrc": "3.1.0",
+    "@eslint/js": "9.13.0",
+    "@rollup/plugin-babel": "6.0.4",
+    "@rollup/plugin-node-resolve": "15.3.0",
+    "babel-jest": "29.7.0",
+    "babel-loader": "9.2.1",
+    "eslint": "9.13.0",
+    "eslint-import-resolver-webpack": "0.13.9",
+    "eslint-plugin-import": "2.31.0",
+    "eslint-plugin-react": "7.37.2",
+    "eslint-plugin-react-hooks": "5.0.0",
+    "globals": "^15.11.0",
+    "jest": "29.7.0",
+    "js-var-type": "0.2.6",
+    "lodash-es": "4.17.21",
+    "prettier": "3.3.3",
+    "prop-types": "15.8.1",
+    "react": "18.3.1",
+    "react-dom": "18.3.1",
+    "rollup": "4.24.3",
+    "standard-version": "9.5.0",
+    "webpack": "5.96.1",
+    "webpack-cli": "5.1.4",
+    "webpack-dev-server": "5.1.0"
   },
   "peerDependencies": {
-    "@babel/runtime": "^7.12.5",
-    "js-var-type": "^0.2.6",
-    "lodash-es": "^4.17.20",
-    "prop-types": "^15.7.2",
-    "react": "^17.0.1",
-    "react-dom": "^17.0.1"
+    "@babel/runtime": "7.26.0",
+    "js-var-type": "0.2.6",
+    "lodash-es": "4.17.21",
+    "prop-types": "15.8.1",
+    "react": "18.3.1",
+    "react-dom": "18.3.1"
   }
 }
