@@ -1,8 +1,1 @@
-/* eslint-disable no-undef */
-/* eslint-disable import/no-commonjs */
-/* eslint-disable import/unambiguous */
-module.exports = {
-  moduleNameMapper: {
-    'js-var-type': 'js-var-type/dist/main.cjs',
-  },
-};
+module.exports = {};
